requirements.txt
setuptools
pygame-ce
